int to uint64_t

+++ src/ghostquest/ghostquest.cpp
@@ -37,7 +37,7 @@ ACTION ghostquest::genchar(name username, asset quantity, int limit) // generate
             new_ghost.status = STANDBY;
             new_ghost.initial_hp = 100 + rng(50);
             gen_stat(new_ghost); // generate status for character
-            game_data.character.insert(game_data.character.end(), pair<int, ghost>(key, new_ghost));
+            game_data.character.insert(game_data.character.end(), pair<uint64_t, ghost>(key, new_ghost));
         }
     });
 }
